% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.R
\name{get_toxprints}
\alias{get_toxprints}
\title{Calculate ToxPrint features}
\usage{
get_toxprints(smiles)
}
\arguments{
\item{smiles}{character of chemical structure representation with SMILES string}
}
\description{
Hidden function to feed a single string to the API and return a named list with
ToxPrints as names and the binary value of each Print as the values .
}
