% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_models.R
\name{in_domain}
\alias{in_domain}
\title{Predict if a record is in a model's domain of applicability}
\usage{
in_domain(df, models, chemical_id = "chemical_id", method = "jaccard")
}
\arguments{
\item{df}{ToxPrint DataFrame of chemicals to predict}

\item{models}{list of models, or "classes", for which to determine domain}

\item{method}{distance metric to use for similarity calculation}

\item{sim_cut}{NULL or float. If NULL, use stored values calculated for a
similarity threshold of 0.5, if float fraction from 0 to 1 for how similar
records should be, default=NULL}
}
\description{
For each record in a testing set of data, applicability_domain
determines if that record lies within the domain of applicability of the
training set for a given class in a classification model. See the following
for more information on these methods:
Rational selection of training and test sets for the development
of valid QSAR models, A. Golbraikh, et al., J. Comp.-Aided Mol. Design,
17 (2003), 241 -- 253.
Predictive QSAR modeling workflow, model applicability domains,
and virtual screening, A. Tropsha and A. Golbraikh, Curr. Pharm. Design,
13 (2007), 3494 -- 3504.
}
