% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_models.R
\name{predict_all_in_domain}
\alias{predict_all_in_domain}
\title{Predict a chemical for all QSURs and only return predictions that are in domain
for each model}
\usage{
predict_all_in_domain(
  models,
  df,
  type = "prob",
  chemical_id = "chemical_id",
  method = "jaccard"
)
}
\arguments{
\item{models}{list of models, or "classes", for which to determine domain}

\item{df}{ToxPrint DataFrame of chemicals to predict}

\item{chemical_id}{name of column in dataframe that is consider the chemical
identifier columns}

\item{method}{distance metric to use for similarity calculation}
}
\description{
Predict a chemical for all QSURs and only return predictions that are in domain
for each model
}
