% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_models.R
\name{predict_all_QSUR}
\alias{predict_all_QSUR}
\title{Predict with all 39 QSUR Models}
\usage{
predict_all_QSUR(models, df, type = "prob", chemical_id = "chemical_id")
}
\arguments{
\item{df}{ToxPrint DataFrame of chemicals to predict}

\item{type}{type of prediction to get from RF model (prob, response, or vote)}

\item{chemical_id}{name of chemical_id used, default is \code{chemical_id}}

\item{model}{randomForest model object}
}
\description{
Provides predictions of all 39 valid harmonized use QSUR model from a passed
data frame of N chemicals with all 730 ToxPrints
}
\examples{
predict_all_QSUR()
}
