% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_models.R
\name{predict_one_QSUR}
\alias{predict_one_QSUR}
\title{Predict with a Single QSUR Model}
\usage{
predict_one_QSUR(model, df, type = "prob", chemical_id = "chemical_id")
}
\arguments{
\item{model}{randomForest model object}

\item{df}{ToxPrint DataFrame of chemicals to predict}

\item{type}{type of prediction to get from RF model (prob, response, or vote)}

\item{chemical_id}{name of chemical_id used, default is \code{chemical_id}}
}
\description{
Provides predictions of a single harmonized use QSUR model from a passed data
frame of N chemicals with all 730 ToxPrints
}
\examples{
predict_one_QSUR()
}
