% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qsur.r
\docType{package}
\name{QSUR-package}
\alias{QSUR}
\alias{QSUR-package}
\title{QSUR.
QSUR: A package storing 39 structure based QSUR models.}
\description{
Description: Encapsulates the quantitative structure-use relationship models developed
in \emph{High-throughput screening of chemicals as functional substitutes using
structure-based classification models} (DOI: https://doi.org/10.1039/c6gc02744j)
along with the capabilities to load new external chemicals for prediction and use
the models themeselves for predicting.
}
\author{
\strong{Maintainer}: Katherine Phillips \email{katherine.phillips@epa.gov} (\href{https://orcid.org/0000-0001-5703-0698}{ORCID})

}
