% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.R
\name{qsur_models}
\alias{qsur_models}
\title{Load QSUR Models}
\usage{
qsur_models()
}
\arguments{
\item{harmonized_use}{name of model to use}
}
\description{
Import a named list of all 39 valid, structure-only QSUR models contained
within this package. Names of the list are the harmonized use predicted by
that randomForest object value in the named list.
}
\examples{
qsur_models()
}
