% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.R
\name{valid_uses}
\alias{valid_uses}
\title{Get valid QSUR model names}
\usage{
valid_uses()
}
\description{
Return a list of all valid structure-only QSUR models contained in package
}
\examples{
valid_uses()
}
