% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.R
\name{qsur_models}
\alias{qsur_models}
\title{Load QSUR Models}
\usage{
qsur_models(harmonized_use)
}
\arguments{
\item{harmonized_use}{name of model to use}
}
\description{
Load list of all or a single QSUR model
}
\examples{
qsur_models()
}
