% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.R
\name{read_toxprints}
\alias{read_toxprints}
\title{Read ToxPrint file for prediction}
\usage{
read_toxprints(io, source, chemical_id = "chemical_id")
}
\arguments{
\item{io}{path and name of file to read}

\item{source}{either \code{chemotyper} or \code{ccd}}

\item{chemical_id}{name of chemical_id used, default is \code{chemical_id}}
}
\description{
Read in ToxPrint file either from the ChemoTyper application or the CCD
}
\examples{
read_toxprintsl()
}
